requirements.txt
groq
sentence-transformers
numpy
python-dotenv
fastapi
uvicorn
